Move JSON / console output methods in named_solution.h to value_proxy.h and string_utility.h.
