+++ cppmh/model/named_solution.h
@@ -39,150 +39,6 @@ class NamedSolution {
     T_Expression m_objective;
     bool         m_is_feasible;
 
-    /*************************************************************************/
-    template <class T_Value>
-    inline static void print_values(
-        const std::unordered_map<std::string, ValueProxy<T_Value>>
-                           a_VALUE_PROXIES,
-        const std::string& a_CATEGORY) {
-        for (const auto& item : a_VALUE_PROXIES) {
-            auto& proxy              = item.second;
-            int   number_of_elements = proxy.number_of_elements();
-            for (auto i = 0; i < number_of_elements; i++) {
-                utility::print(a_CATEGORY + "." + proxy.flat_indexed_names(i) +
-                               " = " +
-                               std::to_string(proxy.flat_indexed_values(i)));
-            }
-        }
-    }
-
-    /*************************************************************************/
-    template <class T_Value>
-    inline static void write_values_by_name(
-        std::ofstream* a_ofs,
-        const std::unordered_map<std::string, ValueProxy<T_Value>>
-                           a_VALUE_PROXIES,
-        const std::string& a_CATEGORY, const int a_INDENT_LEVEL) {
-        int indent_level = a_INDENT_LEVEL;
-
-        *a_ofs << indent_spaces(indent_level) << "\"" + a_CATEGORY + "\" : {"
-               << std::endl;
-        indent_level++;
-
-        int count              = 0;
-        int value_proxies_size = a_VALUE_PROXIES.size();
-        for (const auto& item : a_VALUE_PROXIES) {
-            auto& proxy              = item.second;
-            int   number_of_elements = proxy.number_of_elements();
-            for (auto i = 0; i < number_of_elements; i++) {
-                *a_ofs << indent_spaces(indent_level)
-                       << "\"" + proxy.flat_indexed_names(i) + "\" : " +
-                              std::to_string(proxy.flat_indexed_values(i));
-
-                if ((i == number_of_elements - 1) &&
-                    (count == value_proxies_size - 1)) {
-                    *a_ofs << std::endl;
-                } else {
-                    *a_ofs << "," << std::endl;
-                }
-            }
-            count++;
-        }
-        indent_level--;
-        *a_ofs << indent_spaces(indent_level) << "}," << std::endl;
-    }
-
-    /*************************************************************************/
-    template <class T_Value>
-    inline static void write_values_by_array(
-        std::ofstream* a_ofs,
-        const std::unordered_map<std::string, ValueProxy<T_Value>>
-                           a_VALUE_PROXIES,
-        const std::string& a_CATEGORY, const int a_INDENT_LEVEL) {
-        int indent_level = a_INDENT_LEVEL;
-
-        *a_ofs << indent_spaces(indent_level) << "\"" + a_CATEGORY + "\" : {"
-               << std::endl;
-        indent_level++;
-
-        int count              = 0;
-        int value_proxies_size = a_VALUE_PROXIES.size();
-        for (const auto& item : a_VALUE_PROXIES) {
-            auto& proxy                = item.second;
-            int   number_of_dimensions = proxy.number_of_dimensions();
-            int   number_of_elements   = proxy.number_of_elements();
-
-            *a_ofs << indent_spaces(indent_level)
-                   << "\"" + item.first + "\" : [" << std::endl;
-            indent_level++;
-
-            int current_dimension = 0;
-            for (auto i = 0; i < number_of_elements; i++) {
-                std::vector<int> index = proxy.multi_dimensional_index(i);
-
-                for (auto j = current_dimension; j < number_of_dimensions - 1;
-                     j++) {
-                    if (index[j + 1] == 0) {
-                        *a_ofs << indent_spaces(indent_level) << "["
-                               << std::endl;
-                        indent_level++;
-                        current_dimension++;
-                    } else {
-                        break;
-                    }
-                }
-                if (index[current_dimension] ==
-                    proxy.shape()[current_dimension] - 1) {
-                    *a_ofs << indent_spaces(indent_level)
-                           << std::to_string(proxy.flat_indexed_values(i))
-                           << std::endl;
-                } else {
-                    *a_ofs << indent_spaces(indent_level)
-                           << std::to_string(proxy.flat_indexed_values(i))
-                           << "," << std::endl;
-                }
-
-                for (auto j = current_dimension; j > 0; j--) {
-                    if (index[j] == proxy.shape()[j] - 1) {
-                        indent_level--;
-                        current_dimension--;
-                        if (index[j - 1] == proxy.shape()[j - 1] - 1) {
-                            *a_ofs << indent_spaces(indent_level) << "]"
-                                   << std::endl;
-                        } else {
-                            *a_ofs << indent_spaces(indent_level) << "],"
-                                   << std::endl;
-                        }
-                    } else {
-                        break;
-                    }
-                }
-            }
-
-            indent_level--;
-            count++;
-
-            if (count == value_proxies_size) {
-                *a_ofs << indent_spaces(indent_level) << "]" << std::endl;
-            } else {
-                *a_ofs << indent_spaces(indent_level) << "]," << std::endl;
-            }
-        }
-
-        indent_level--;
-        *a_ofs << indent_spaces(indent_level) << "}," << std::endl;
-    }
-
-    /*************************************************************************/
-    inline static std::string indent_spaces(const int a_INDENT_LEVEL) {
-        std::string indent_width  = "    ";
-        std::string indent_spaces = "";
-        for (auto i = 0; i < a_INDENT_LEVEL; i++) {
-            indent_spaces += indent_width;
-        }
-        return indent_spaces;
-    }
-
    public:
     /*************************************************************************/
     NamedSolution(void) {
@@ -202,22 +58,22 @@ class NamedSolution {
 
     /*************************************************************************/
     inline void print_variable_values(void) const {
-        print_values(m_variable_value_proxies, "variable");
+        model::print_values(m_variable_value_proxies, "variable");
     }
 
     /*************************************************************************/
     inline void print_expression_values(void) const {
-        print_values(m_expression_value_proxies, "expression");
+        model::print_values(m_expression_value_proxies, "expression");
     }
 
     /*************************************************************************/
     inline void print_constraint_values(void) const {
-        print_values(m_constraint_value_proxies, "constraint");
+        model::print_values(m_constraint_value_proxies, "constraint");
     }
 
     /*************************************************************************/
     inline void print_violation_values(void) const {
-        print_values(m_violation_value_proxies, "violation");
+        model::print_values(m_violation_value_proxies, "violation");
     }
 
     /*************************************************************************/
@@ -225,34 +81,34 @@ class NamedSolution {
         int indent_level = 0;
 
         std::ofstream ofs(a_FILE_NAME.c_str());
-        ofs << indent_spaces(indent_level) << "{" << std::endl;
+        ofs << utility::indent_spaces(indent_level) << "{" << std::endl;
         indent_level++;
 
         /// Decision variables
-        write_values_by_name(&ofs, m_variable_value_proxies, "variables",
-                             indent_level);
+        model::write_values_by_name(&ofs, m_variable_value_proxies,  //
+                                    "variables", indent_level);
 
         /// Expressions
-        write_values_by_name(&ofs, m_expression_value_proxies, "expressions",
-                             indent_level);
+        model::write_values_by_name(&ofs, m_expression_value_proxies,
+                                    "expressions", indent_level);
 
         /// Constraints
-        write_values_by_name(&ofs, m_constraint_value_proxies, "constraints",
-                             indent_level);
+        model::write_values_by_name(&ofs, m_constraint_value_proxies,
+                                    "constraints", indent_level);
 
         /// Violations
-        write_values_by_name(&ofs, m_violation_value_proxies, "violations",
-                             indent_level);
+        model::write_values_by_name(&ofs, m_violation_value_proxies,
+                                    "violations", indent_level);
 
         /// Others
-        ofs << indent_spaces(indent_level)
+        ofs << utility::indent_spaces(indent_level)
             << "\"is_feasible\" : " << (m_is_feasible ? "true," : "false,")
             << std::endl;
-        ofs << indent_spaces(indent_level)
+        ofs << utility::indent_spaces(indent_level)
             << "\"objective\" : " + std::to_string(m_objective) << std::endl;
 
         indent_level--;
-        ofs << indent_spaces(indent_level) << "}" << std::endl;
+        ofs << utility::indent_spaces(indent_level) << "}" << std::endl;
         ofs.close();
     }
 
@@ -261,34 +117,34 @@ class NamedSolution {
         int indent_level = 0;
 
         std::ofstream ofs(a_FILE_NAME.c_str());
-        ofs << indent_spaces(indent_level) << "{" << std::endl;
+        ofs << utility::indent_spaces(indent_level) << "{" << std::endl;
         indent_level++;
 
         /// Decision variables
-        write_values_by_array(&ofs, m_variable_value_proxies, "variables",
-                              indent_level);
+        model::write_values_by_array(&ofs, m_variable_value_proxies,
+                                     "variables", indent_level);
 
         /// Expressions
-        write_values_by_array(&ofs, m_expression_value_proxies, "expressions",
-                              indent_level);
+        model::write_values_by_array(&ofs, m_expression_value_proxies,
+                                     "expressions", indent_level);
 
         /// Constraints
-        write_values_by_array(&ofs, m_constraint_value_proxies, "constraints",
-                              indent_level);
+        model::write_values_by_array(&ofs, m_constraint_value_proxies,
+                                     "constraints", indent_level);
 
         /// Violations
-        write_values_by_array(&ofs, m_violation_value_proxies, "violations",
-                              indent_level);
+        model::write_values_by_array(&ofs, m_violation_value_proxies,
+                                     "violations", indent_level);
 
         /// Others
-        ofs << indent_spaces(indent_level)
+        ofs << utility::indent_spaces(indent_level)
             << "\"is_feasible\" : " << (m_is_feasible ? "true," : "false,")
             << std::endl;
-        ofs << indent_spaces(indent_level)
+        ofs << utility::indent_spaces(indent_level)
             << "\"objective\" : " + std::to_string(m_objective) << std::endl;
 
         indent_level--;
-        ofs << indent_spaces(indent_level) << "}" << std::endl;
+        ofs << utility::indent_spaces(indent_level) << "}" << std::endl;
         ofs.close();
     }
