+++ cppmh/model/solution_score.h
@@ -43,7 +43,8 @@ struct SolutionScore {
         /// nothing to do
     }
 };
-/*************************************************************************/
+
+/*****************************************************************************/
 inline int argmax_index_objective_improvement(
     const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> objective_improvements(a_SCORES.size());
@@ -56,7 +57,7 @@ inline int argmax_index_objective_improvement(
     return utility::argmax(objective_improvements);
 }
 
-/*************************************************************************/
+/*****************************************************************************/
 inline int argmin_index_objective(const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> objectives(a_SCORES.size());
 
@@ -67,7 +68,7 @@ inline int argmin_index_objective(const std::vector<SolutionScore> &a_SCORES) {
     return utility::argmin(objectives);
 }
 
-/*************************************************************************/
+/*****************************************************************************/
 inline int argmin_index_local_penalty(
     const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> local_penalties(a_SCORES.size());
@@ -79,7 +80,7 @@ inline int argmin_index_local_penalty(
     return utility::argmin(local_penalties);
 }
 
-/*************************************************************************/
+/*****************************************************************************/
 inline int argmin_index_global_penalty(
     const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> global_penalties(a_SCORES.size());
@@ -91,7 +92,7 @@ inline int argmin_index_global_penalty(
     return utility::argmin(global_penalties);
 }
 
-/*************************************************************************/
+/*****************************************************************************/
 inline int argmin_index_local_augmented_objective(
     const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> local_augmented_objectives(a_SCORES.size());
@@ -104,7 +105,7 @@ inline int argmin_index_local_augmented_objective(
     return utility::argmin(local_augmented_objectives);
 }
 
-/*************************************************************************/
+/*****************************************************************************/
 inline int argmin_index_global_augmented_objective(
     const std::vector<SolutionScore> &a_SCORES) {
     std::vector<double> global_augmented_objectives(a_SCORES.size());
